COP-10471: Addressed comments on PR
